# docs/swagger.yaml
basePath: /api/v1
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://www.example.com/support
  description: 这是一个用户注册和认证系统的后端API
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Backend API
  version: "1.0"
paths:
...
  /friends/{friend_id}:
    delete:
      parameters:
      - description: 好友ID
        in: path
        name: friend_id
        required: true
        type: string
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: 删除好友
      tags:
      - 好友
  /friends/blocks:
    get:
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: 黑名单列表
      tags:
      - 好友
  /friends/blocks/{user_id}:
    delete:
      parameters:
      - description: 用户ID
        in: path
        name: user_id
        required: true
        type: string
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: 取消拉黑
      tags:
      - 好友
    post:
      parameters:
      - description: 用户ID
        in: path
        name: user_id
        required: true
        type: string
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: 拉黑用户
      tags:
      - 好友
  /friends/list:
    get:
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: 获取好友列表
      tags:
      - 好友
  /friends/remarks/{friend_id}:
    patch:
      consumes:
      - application/json
      parameters:
      - description: 好友ID
        in: path
        name: friend_id
        required: true
        type: string
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: 更新好友备注
      tags:
      - 好友
  /friends/requests:
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: 发起好友请求
      tags:
      - 好友
  /friends/requests/{id}:
    delete:
      parameters:
      - description: 请求ID
        in: path
        name: id
        required: true
        type: string
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: 撤回好友请求
      tags:
      - 好友
  /friends/requests/{id}/accept:
    post:
      consumes:
      - application/json
      parameters:
      - description: 请求ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: 接受好友请求
      tags:
      - 好友
  /friends/requests/{id}/reject:
    post:
      consumes:
      - application/json
      parameters:
      - description: 请求ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: 拒绝好友请求
      tags:
      - 好友
  /friends/requests/incoming:
    get:
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: 入站好友请求列表
      tags:
      - 好友
  /friends/requests/outgoing:
    get:
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: 出站好友请求列表
      tags:
      - 好友
...
  /ws/chat:
    get:
      description: '通过 WebSocket 建立一对一聊天连接。鉴权方式：优先读取 HTTP Header `Authorization: Bearer
        <access_token>`；浏览器无法自定义 Header 时可使用 query 参数 `?token=<access_token>`。'
      parameters:
      - description: Access Token（可选，浏览器场景使用）
        in: query
        name: token
        type: string
      produces:
      - application/json
      responses:
        "101":
          description: Switching Protocols
          schema:
            type: string
      summary: WebSocket 聊天连接
      tags:
      - chat
securityDefinitions:
  ApiKeyAuth:
    description: Description for what is this security definition being used
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"